** added a redraw routine to gridvis to ** redraw everything on purpose ** had to add an ActionHistory for that (just a vector) ** this history can be emptied with clear_window (in eclipse) ** added according predicates to interface-pl ** replaced refresh with redraw in update (exogf_Update)
stf


git-svn-id: https://svn.kbsg.rwth-aachen.de/repos/kbsgolog@40 d9e49986-696a-4e79-9435-a9269e428f0d

+++ utils/gridvis/graphics.cpp
@@ -541,7 +541,7 @@ int mouse_coords (int *x, int *y, int *button, jwin *w)
       break;
     }
   if (valid && (*x >= 0) && (*y >= 0)
-            && (*x < ww) && (*y < wh))
+      && ((unsigned int)*x < ww) && ((unsigned int)*y < wh))
     return(1);
   else
     return(0);
